+++ doc/CHANGES.rst
@@ -25,6 +25,7 @@ CHANGELOG
   **error_found** in the result of the import
 * [FIX] ``Model.get_primary_keys`` with polymorphism (single table)
 * Remove ``column.uText``, ``column.uString``, ``column.SmallInteger``
+* [ADD] ``column.PhoneNumber``
 
 0.17.2 (2018-02-27)
 -------------------